requirements-airflow.txt
psycopg2-binary==2.9.10
pandas==2.2.3
minio==7.2.15
pyarrow==18.1.0
